Create separated modules for DOM manipulation and localStorage handling
