+++ index.html
@@ -28,6 +28,10 @@
         <option value="normal">Normal</option>
         <option value="high">High</option>
       </select>
+      <select id="status-list" name="status-list">
+        <option value="unfinished">Unfinished</option>
+        <option value="finished">Finished</option>
+      </select>
 
       <input type="submit" id="submit-activity" value="Add">
     </form>
@@ -45,6 +49,7 @@
         <th>Due Date</th>
         <th>Description</th>
         <th>Priority</th>
+        <th>Status</th>
         <th>Edit</th>
       </tr>
     </table>
